refactor: extract DebugBundleTypes.swift from DebugBundleExporter.swift
Extracts 3 public support types (DebugInfoSnapshot, DebugSessionMetadata,
DebugBundleSnapshot) from DebugBundleExporter.swift into a new DebugBundleTypes.swift.

Byte-identical move; SHA-256 7700572e83dbecc8c71f92faf1b9f9fe5fcb46d8d02b4cdc27940266ca51010d
verified against origin/main:DebugBundleExporter.swift lines 5-163.

Reduces DebugBundleExporter.swift from 276 to 117 lines. The primary
DebugBundleExporter struct + private DebugSystemInfoDocument helper + extension
remain in the exporter file.

Tests: 667 (unchanged).
